% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recognize.R
\docType{package}
\name{abbyyR-package}
\alias{abbyyR}
\alias{abbyyR-package}
\title{abbyyR: R Client for the Abbyy Cloud OCR}
\description{
Compare OCR documents to gold standard transcriptions. Takes two folders
}
\author{
Gaurav Sood
}

