% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_all.R
\name{compare_all}
\alias{compare_all}
\title{Compare All}
\usage{
compare_all(ocr_dir = NULL, gold_dir = NULL, remove_extra_space = TRUE,
  normalize = TRUE)
}
\arguments{
\item{ocr_dir}{path to directory containing files with OCRd text; required}

\item{gold_dir}{path to directory containing files with human transcribed text; required}

\item{remove_extra_space}{a dummy indicating whether or not extra spaces should be removed from the OCR file; default is TRUE}

\item{normalize}{normalize string distance measures --- otherwise longer the document, greater the distance}
}
\value{
error rate
}
\description{
Takes path to directory with OCRd files and path to directory with human transcribed files and calculates error rate.
Calls compare_txt
}
\examples{
\dontrun{
compare_all(ocr_dir="path_to_ocr_dir", gold_dir="path_to_gold_dir", 
	           remove_extra_space=TRUE)
}
}

