% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recognize.R
\docType{package}
\name{recognize-package}
\alias{recognize}
\alias{recognize-package}
\title{recognize: Assess OCR quality}
\description{
Compare OCR documents to gold standard transcriptions. Takes two folders
}
\author{
Gaurav Sood
}

