% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_docs.R
\name{sample_docs}
\alias{sample_docs}
\title{Sample Docs}
\usage{
sample_docs(ocr_dir = NULL, output_dir = NULL, nsamp = NULL)
}
\arguments{
\item{ocr_dir}{path to folder containing OCRd documents; required}

\item{output_dir}{path to folder containing sampled OCRd documents; required}

\item{nsamp}{number of sampled files; required}
}
\value{
a vector of names of the sampled files. It also creates a directory with sampled files
}
\description{
Output a weighted random sample of documents. Documents weighted by their length.
}
\examples{
\dontrun{
sample_docs(ocr_dir="path_to_ocr_dir", output_dir = "path_to_output_dir_with_sample_docs")
}
}

